# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CourtBook API - Court booking platform for Lithuania
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/create-checkout:
    post:
      operationId: createCheckoutSession
      tags: [payments]
      summary: Create a Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutBody"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
  /payments/confirm:
    post:
      operationId: confirmPayment
      tags: [payments]
      summary: Confirm payment and finalize booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPaymentBody"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        courtId:
          type: integer
        courtName:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        stripeSessionId:
          type: string
        refundAmount:
          type: number
        stripeRefundId:
          type: string
        managementToken:
          type: string
          nullable: true
          description: Opaque token for guest bookings to manage the reservation without an account.
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - courtId
        - customerName
        - customerEmail
        - date
        - startTime
        - endTime
        - totalPrice
        - status
        - createdAt
...
    CreateCheckoutBody:
      type: object
      properties:
        bookingId:
          type: integer
        successUrl:
          type: string
        cancelUrl:
          type: string
      required:
        - bookingId
        - successUrl
        - cancelUrl
    CheckoutSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
        url:
          type: string
      required:
        - sessionId
        - url
    ConfirmPaymentBody:
      type: object
      properties:
        sessionId:
          type: string
      required:
        - sessionId
...